+++ common/src/main/java/me/pandamods/fallingtrees/config/CommonConfig.java
@@ -4,11 +4,11 @@ import me.pandamods.fallingtrees.FallingTrees;
 import me.pandamods.fallingtrees.config.common.FeaturesConfig;
 import me.pandamods.fallingtrees.config.common.FilterConfig;
 import me.pandamods.fallingtrees.config.common.LimitationsConfig;
-import me.shedaniel.autoconfig.ConfigData;
-import me.shedaniel.autoconfig.annotation.Config;
-import me.shedaniel.autoconfig.annotation.ConfigEntry;
+import me.pandamods.pandalib.PandaLib;
+import me.pandamods.pandalib.config.api.Config;
+import me.pandamods.pandalib.config.api.ConfigData;
 
-@Config(name = FallingTrees.MOD_ID + "_common")
+@Config(name = FallingTrees.MOD_ID + "_common", modId = PandaLib.MOD_ID, synchronize = true)
 public class CommonConfig implements ConfigData {
 	public boolean isCrouchMiningAllowed = true;
 	public boolean multiplyToolDamage = true;
@@ -16,15 +16,9 @@ public class CommonConfig implements ConfigData {
 
 	public float treeLifetimeLength = 4;
 
-	@ConfigEntry.Category("filter")
-	@ConfigEntry.Gui.TransitiveObject
 	public FilterConfig filter = new FilterConfig();
 
-	@ConfigEntry.Category("limitations")
-	@ConfigEntry.Gui.TransitiveObject
 	public LimitationsConfig limitations = new LimitationsConfig();
 
-	@ConfigEntry.Category("features")
-	@ConfigEntry.Gui.TransitiveObject
 	public FeaturesConfig features = new FeaturesConfig();
 }