+++ common/src/main/java/me/pandamods/fallingtrees/config/FallingTreesConfig.java
@@ -1,52 +1,27 @@
 package me.pandamods.fallingtrees.config;
 
-import dev.architectury.platform.Platform;
 import me.pandamods.fallingtrees.FallingTrees;
-import me.pandamods.fallingtrees.config.screen.ConfigScreen;
-import me.pandamods.fallingtrees.event.EventHandler;
-import me.pandamods.fallingtrees.network.ConfigPacket;
-import me.shedaniel.autoconfig.AutoConfig;
-import me.shedaniel.autoconfig.ConfigHolder;
-import me.shedaniel.autoconfig.serializer.GsonConfigSerializer;
-import me.shedaniel.autoconfig.serializer.PartitioningSerializer;
-import net.fabricmc.api.EnvType;
-import net.minecraft.client.Minecraft;
-import net.minecraft.client.gui.screens.Screen;
-import net.minecraft.world.InteractionResult;
+import me.pandamods.pandalib.config.api.ConfigRegistry;
+import me.pandamods.pandalib.config.api.holders.ClientConfigHolder;
+import me.pandamods.pandalib.config.api.holders.CommonConfigHolder;
+import net.minecraft.world.entity.player.Player;
 
 public class FallingTreesConfig {
-	public final ConfigHolder<ClientConfig> clientConfigHolder;
-	public final ConfigHolder<CommonConfig> commonConfigHolder;
-	private CommonConfig commonConfig;
+	public final ClientConfigHolder<ClientConfig> clientConfigHolder = ConfigRegistry.registerClient(ClientConfig.class);
+	public final CommonConfigHolder<CommonConfig> commonConfigHolder = ConfigRegistry.registerCommon(CommonConfig.class);
 
 	public FallingTreesConfig() {
-		clientConfigHolder = AutoConfig.register(ClientConfig.class, GsonConfigSerializer::new);
-		commonConfigHolder = AutoConfig.register(CommonConfig.class, GsonConfigSerializer::new);
-
-		if (Platform.getEnv().equals(EnvType.CLIENT)) {
-			clientConfigHolder.registerSaveListener(this::saveConfig);
-		}
 	}
 
-	private InteractionResult saveConfig(ConfigHolder<ClientConfig> configHolder, ClientConfig config) {
-		if (Minecraft.getInstance().level != null) {
-			ConfigPacket.sendToServer();
-		}
-
-		return InteractionResult.PASS;
+	public static ClientConfig getClientConfig(Player player) {
+		return FallingTrees.CONFIG.clientConfigHolder.getClient(player);
 	}
 
 	public static ClientConfig getClientConfig() {
-		return FallingTrees.CONFIG.clientConfigHolder.getConfig();
+		return FallingTrees.CONFIG.clientConfigHolder.get();
 	}
 
 	public static CommonConfig getCommonConfig() {
-		if (FallingTrees.CONFIG.commonConfig != null)
-			return FallingTrees.CONFIG.commonConfig;
-		return FallingTrees.CONFIG.commonConfigHolder.getConfig();
-	}
-
-	public void setCommonConfig(CommonConfig commonConfig) {
-		this.commonConfig = commonConfig;
+		return FallingTrees.CONFIG.commonConfigHolder.get();
 	}
 }