+++ common/src/main/java/me/pandamods/fallingtrees/event/EventHandler.java
@@ -2,9 +2,7 @@ package me.pandamods.fallingtrees.event;
 
 import dev.architectury.event.EventResult;
 import dev.architectury.event.events.client.ClientLifecycleEvent;
-import dev.architectury.event.events.client.ClientPlayerEvent;
 import dev.architectury.event.events.common.BlockEvent;
-import dev.architectury.event.events.common.PlayerEvent;
 import dev.architectury.platform.Platform;
 import dev.architectury.registry.client.level.entity.EntityRendererRegistry;
 import dev.architectury.utils.value.IntValue;
@@ -14,11 +12,9 @@ import me.pandamods.fallingtrees.client.render.TreeRenderer;
 import me.pandamods.fallingtrees.config.CommonConfig;
 import me.pandamods.fallingtrees.config.FallingTreesConfig;
 import me.pandamods.fallingtrees.entity.TreeEntity;
-import me.pandamods.fallingtrees.network.ConfigPacket;
 import me.pandamods.fallingtrees.registry.EntityRegistry;
 import net.fabricmc.api.EnvType;
 import net.minecraft.client.Minecraft;
-import net.minecraft.client.player.LocalPlayer;
 import net.minecraft.core.BlockPos;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.stats.Stats;
@@ -34,18 +30,7 @@ import java.util.Set;
 
 public class EventHandler {
 	public static void register() {
-		if (Platform.getEnv() == EnvType.CLIENT) {
-			if (Platform.isFabric()) ClientLifecycleEvent.CLIENT_SETUP.register(EventHandler::onClientSetup);
-			if (Platform.isForge()) onClientSetup(Minecraft.getInstance());
-		}
 		BlockEvent.BREAK.register(EventHandler::onBlockBreak);
-		PlayerEvent.PLAYER_JOIN.register(EventHandler::onPlayerJoin);
-	}
-
-	private static void onClientSetup(Minecraft minecraft) {
-		ClientPlayerEvent.CLIENT_PLAYER_JOIN.register(EventHandler::onClientPlayerJoin);
-
-		EntityRendererRegistry.register(EntityRegistry.TREE, TreeRenderer::new);
 	}
 
 	private static EventResult onBlockBreak(Level level, BlockPos blockPos, BlockState blockState, ServerPlayer serverPlayer, IntValue intValue) {
@@ -53,13 +38,6 @@ public class EventHandler {
 			return EventResult.interruptFalse();
 		}
 		return EventResult.pass();
-	}
-
-	private static void onPlayerJoin(ServerPlayer serverPlayer) {
-		ConfigPacket.sendToPlayer(serverPlayer);
-	}
-	private static void onClientPlayerJoin(LocalPlayer localPlayer) {
-		ConfigPacket.sendToServer();
 	}
 
 	public static boolean makeTreeFall(BlockPos blockPos, LevelAccessor level, Player player) {