+++ common/src/main/java/me/pandamods/fallingtrees/mixin/MultiPlayerGameModeMixin.java
@@ -1,0 +1,9 @@
+package me.pandamods.fallingtrees.mixin;
+
+import net.minecraft.client.multiplayer.MultiPlayerGameMode;
+import org.spongepowered.asm.mixin.Mixin;
+
+@Mixin(MultiPlayerGameMode.class)
+public class MultiPlayerGameModeMixin {
+
+}