+++ common/src/main/java/me/pandamods/fallingtrees/mixin/ServerPlayerGameModeMixin.java
@@ -1,0 +1,9 @@
+package me.pandamods.fallingtrees.mixin;
+
+import net.minecraft.server.level.ServerPlayerGameMode;
+import org.spongepowered.asm.mixin.Mixin;
+
+@Mixin(ServerPlayerGameMode.class)
+public class ServerPlayerGameModeMixin {
+
+}