+++ common/src/main/java/me/pandamods/fallingtrees/trees/CactusTree.java
@@ -3,12 +3,10 @@ package me.pandamods.fallingtrees.trees;
 import me.pandamods.fallingtrees.api.TreeType;
 import me.pandamods.fallingtrees.config.FallingTreesConfig;
 import net.minecraft.core.BlockPos;
-import net.minecraft.core.Direction;
 import net.minecraft.world.level.LevelAccessor;
 import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.state.BlockState;
 
-import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.Set;