+++ fabric/src/main/java/me/pandamods/fallingtrees/fabric/ModMenuIntegration.java
@@ -1,17 +1,13 @@
 package me.pandamods.fallingtrees.fabric;
 
-import com.terraformersmc.modmenu.api.ConfigScreenFactory;
 import com.terraformersmc.modmenu.api.ModMenuApi;
-import me.pandamods.fallingtrees.config.FallingTreesConfig;
-import me.pandamods.fallingtrees.config.screen.ConfigScreen;
-import me.shedaniel.autoconfig.AutoConfig;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 
 @Environment(EnvType.CLIENT)
 public class ModMenuIntegration implements ModMenuApi {
-	@Override
-	public ConfigScreenFactory<?> getModConfigScreenFactory() {
-		return ConfigScreen::new;
-	}
+//	@Override
+//	public ConfigScreenFactory<?> getModConfigScreenFactory() {
+//		return ConfigScreen::new;
+//	}
 }