+++ forge/src/main/java/me/pandamods/fallingtrees/forge/FallingTreesForgeClient.java
@@ -3,18 +3,14 @@ package me.pandamods.fallingtrees.forge;
 import me.pandamods.fallingtrees.FallingTrees;
 import me.pandamods.fallingtrees.FallingTreesClient;
 import net.minecraftforge.api.distmarker.Dist;
-import net.minecraftforge.client.ConfigScreenHandler;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
-import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
 
 @Mod.EventBusSubscriber(modid = FallingTrees.MOD_ID, value = Dist.CLIENT, bus = Mod.EventBusSubscriber.Bus.MOD)
-public class FallingTreesClientForge {
+public class FallingTreesForgeClient {
 	@SubscribeEvent
 	public static void clientInit(FMLClientSetupEvent event) {
 		FallingTreesClient.init();
-//		ModLoadingContext.get().registerExtensionPoint(ConfigScreenHandler.ConfigScreenFactory.class, () ->
-//				new ConfigScreenHandler.ConfigScreenFactory((minecraft, screen) -> new ConfigScreen(screen)));
 	}
 }